Change directory name from sidekiq-ecs-scaler to sidekiq_ecs_scaler
In Ruby, it is common to use snake cases for file names and directory names.
The only exception is `lib/sidekiq-ecs-scaler.rb`.
Because I want the gem name to be a kebab case.
