+++ Dockerfile
@@ -7,8 +7,8 @@ RUN apt-get update \
 
 WORKDIR /usr/src/sidekiq-ecs-scaler
 
-RUN mkdir -p lib/sidekiq-ecs-scaler \
-    && echo "module SidekiqEcsScaler\n  VERSION = \"0.1.0\"\nend\n" > lib/sidekiq-ecs-scaler/version.rb
+RUN mkdir -p lib/sidekiq_ecs_scaler \
+    && echo "module SidekiqEcsScaler\n  VERSION = \"0.1.0\"\nend\n" > lib/sidekiq_ecs_scaler/version.rb
 
 COPY bin/setup ./bin/
 COPY Gemfile Gemfile.lock sidekiq-ecs-scaler.gemspec .
