+++ lib/sidekiq_ecs_scaler/client.rb
