+++ lib/sidekiq_ecs_scaler/configuration.rb
