+++ lib/sidekiq_ecs_scaler/task_meta_v4.rb
