+++ lib/sidekiq_ecs_scaler/version.rb
