+++ lib/sidekiq_ecs_scaler/worker.rb
