+++ spec/sidekiq_ecs_scaler/client_spec.rb
