+++ spec/sidekiq_ecs_scaler/configuration_spec.rb
