+++ spec/sidekiq_ecs_scaler/task_meta_v4_spec.rb
