+++ spec/sidekiq_ecs_scaler/worker_spec.rb
