+++ spec/sidekiq_ecs_scaler_spec.rb
